+++ src/main/java/chess/domain/Team.java
@@ -1,4 +1,4 @@
-package chess.domain.pieces;
+package chess.domain;
 
 public enum Team {
     BLACK,
